fix(wasm): split node-safe tests out of the browser-tagged test file
wasm_test.go was tagged browser wholesale, which also gated out its
node-safe tests (GetCapturedOutput priority, SplitArgs, buffer
thread-safety, env helpers). Move the browser-only JS-bridge tests to
wasm_browser_test.go and keep wasm_test.go running under the default
js,wasm CI step with no -run allowlist.
